Minimal(ish) repro for typescript@5.4.0-dev.20231222 issue
This shows a failure mode around a package which (oddly!) publishes an
`index.ts` file in its root but also specifies both the `types` and
`exports.types` keys in `package.json` as pointing to an emitted `.d.ts`
file in a `dist` directory. If the `index.ts` is removed or renamed,
`tsc` correctly uses `types` to resolve the location.

This appears to have been introduced by microsoft/TypeScript#55015.